place new button in new location
