+++ src/components/Personal-Zone/Publisher/empty-display/EmptyDisplayStyles.ts
@@ -1,7 +1,30 @@
 import { makeStyles, createStyles } from '@material-ui/core/styles';
+import { DARK_TEXT_COLOR } from '../../../../assets/projectJSS/Colors';
 
 const useStyles = makeStyles((theme) =>
-    createStyles({})
+    createStyles({
+        root: {
+            display: 'flex',
+            flexDirection: 'column',
+            width: '26vw',
+            height: '30vh',
+            margin: 'auto',
+            marginTop: '7vh',
+            justifyContent: 'center',
+            textAlign: 'center'
+        },
+        messageTitle: {
+            fontSize: 'xx-large',
+            fontWeight: 'bolder',
+            color: DARK_TEXT_COLOR,
+            marginBottom: '1.5vh'
+        },
+        messageBody: {
+            fontSize: 'x-large',
+            fontWeight: 600,
+            marginBottom: '1vh'
+        }
+    })
 );
 
 export default useStyles;
