+++ src/components/Personal-Zone/Publisher/header/Header.tsx
@@ -1,47 +1,12 @@
 import * as React from 'react';
-import { useState, useEffect, useRef } from 'react';
 
-import Button from '@material-ui/core/Button';
 import Typography from '@material-ui/core/Typography';
 import Avatar from '@material-ui/core/Avatar';
 
 import styles from './HeaderStyle';
-import PostNewJob from './PostNewJob/PostNewJob';
 
-import { getAllSelectOptions } from '../../../../server/ads';
-import { AllSelectOptions } from '../../../../types/AllSelectOptions';
-
-interface HeaderOwnProps {
-    fetchAllAdsAfterPost: () => void;
-}
-
-const Header: React.FC<HeaderOwnProps> = (props): JSX.Element => {
-    const { fetchAllAdsAfterPost } = props;
+const Header: React.FC = (): JSX.Element => {
     const classes = styles({});
-
-    const [allSelectOptions, setAllSelectOptions] = useState<AllSelectOptions | null>(null);
-    const [openAddDialog, setOpenAddDialog] = useState<boolean>(true/*TODO:false*/);
-    const addButtonRef = useRef(null);
-
-    // when closing dialog - blur the add button
-    useEffect(() => {
-        if (!openAddDialog && addButtonRef) {
-            // @ts-ignore
-            addButtonRef.current.blur();
-        }
-    }, [openAddDialog]);
-
-    useEffect(() => {
-        if (openAddDialog) {
-            getAllSelectOptions()
-              .then(allSelectOptions => setAllSelectOptions(allSelectOptions));
-        }
-    }, [openAddDialog]);
-    
-
-    const onCloseDialog = (): void => {
-        setOpenAddDialog(false);
-    }
 
     const headerTitle = (): JSX.Element => {
         const FIRST_NAME = 'נוי'; // TODO: Extract from mobx, the logged user's first name
@@ -63,39 +28,10 @@ const Header: React.FC<HeaderOwnProps> = (props): JSX.Element => {
         );
     }
 
-    const addNewPostButton = (): JSX.Element => {
-        return (
-            <Button
-                className={classes.addNewPostButton}
-                variant="contained"
-                onClick={() => setOpenAddDialog(true)}
-                ref={addButtonRef}
-            >
-                <Typography
-                    variant='h5'
-                >
-                    + פרסום תפקיד חדש
-                </Typography>
-            </Button>
-        );
-    }
-
     return (
-        <>
-            <div className={classes.root}>           
-                {headerTitle()}
-                {addNewPostButton()}
-                { openAddDialog && allSelectOptions && <PostNewJob 
-                    allSelectOptions={allSelectOptions}
-                    fetchAllAdsAfterPost={fetchAllAdsAfterPost}
-                    closeDialog={onCloseDialog} />}
-            </div>
-            <Typography 
-                variant='h4'
-                className={classes.jobsHeaderTitle}>
-                התעניינו בתפקידים שלך
-            </Typography>
-        </>
+        <div className={classes.root}>           
+            {headerTitle()}
+        </div>
     );
 }
