+++ src/components/Personal-Zone/Publisher/header/HeaderStyle.ts
@@ -14,11 +14,6 @@ const styles = makeStyles({
         display: 'flex',
         flexDirection: 'row'
     },
-    jobsHeaderTitle: {
-        fontWeight: 'bolder',
-        marginTop: '4vh',
-        marginRight: '5vw'
-    },
     avatar: {
         backgroundColor: AVATAR_COLOR,
         margin: '0 1ch',
@@ -26,18 +21,7 @@ const styles = makeStyles({
         width: '6vh',
         fontSize: '3vh'
     },
-    addNewPostButton: {
-        color: LIGHT_COLOR_TEXT,
-        backgroundColor: NEW_JOB_COLOR,
-        "&:hover": {
-            backgroundColor: `#00d699`,
-        },
-        "&:focus": {
-            backgroundColor: `#00eba8`,
-        },
-        minWidth: '10ch',
-        auto: 'left'
-    },
+
     searchBar: {
         width: '50vw',
         fontSize: '1.3em',
